Add files via upload
This second commit adds instructions at the beginning of the game to inform the player about the positions on the board and corresponding numbers. The board is first printed with the numbers 1-9 in positions corresponding to the key in the dictionary. Then there is a function which resets the board, so each position is blank. This function may get reused at the end of the game if a "replay" option is added. The code also has a player_move function, which takes input from the player as a value between 1-9, and prints their mark on the board in that position.