Add files via upload
This third commit adds a counter which will be updated (+1) after each turn. There is also a function to determine who wins using the first_turn and count variables. If the player went first and the count is not an even number, then the player moved last. This function will be used to determine who wins and will be called if a winning state is reached.