Add files via upload
I have added a game_play function which controls each game. I may include an option for it to be called again at the end of the game if the player wants to replay. The function resets the board and counter, calls the player moves, AI moves (function does not yet exist) and the checks for a winner. After each move the board is updated and printed. If there is a winner, this is printed and currently the program ends when this happens.