Add files via upload
I have updated the game_play function to declare when there is a draw. I have also changed the player_move function to include a try and except block (for input validation) and removed the line of code that was printing the board after each invalid input. Now the user will be prompted for input until they enter a number between 1-9 that matches a blank position on the board.